Split OpenCVTypeAdapters
